Improve error checking JIRA SB-727
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@48598 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ scripts/projects/blast/post_build/macosx/ncbi-blast.sh
@@ -17,10 +17,24 @@ ALL_BINS="$BLAST_BINS $MASKING_BINS $DB_BINS"
 
 rm -rf $PRODUCT.dmg $PRODUCT _stage $INSTALLDIR/installer
 mkdir -p _stage/usr/local/ncbi/blast/bin _stage/usr/local/ncbi/blast/doc _stage/private/etc/paths.d
+if [ $? -ne 0 ]; then
+    echo FAILURE
+    exit 1;
+fi
 
-ftp -o _stage/usr/local/ncbi/blast/doc/user_manual.pdf "http://web.ncbi.nlm.nih.gov/bookshelf/picrender.fcgi?book=helpblast&&partid=1763&blobtype=pdf"
+cat > _stage/usr/local/ncbi/blast/doc/README.txt <<EOF
+Documentation available in http://www.ncbi.nlm.nih.gov/books/NBK1762
+EOF
+if [ $? -ne 0 ]; then
+    echo FAILURE
+    exit 1;
+fi
 
 cp -p $SCRIPTDIR/ncbi_blast _stage/private/etc/paths.d
+if [ $? -ne 0 ]; then
+    echo FAILURE
+    exit 1;
+fi
 
 # This is needed because the binary ncbi-blast.pmproj has this string hard
 # coded
@@ -28,26 +42,54 @@ cp -p $INSTALLDIR/LICENSE ./license.txt
 for f in uninstall_ncbi_blast.zip large-Blue_ncbi_logo.tiff ncbi-blast.pmdoc welcome.txt; do
     echo copying $f to local directory
     cp -rp $SCRIPTDIR/$f .
+    if [ $? -ne 0 ]; then
+        echo FAILURE
+        exit 1;
+    fi
 done
 
 for bin in $ALL_BINS; do
     echo copying $bin
     cp -p $INSTALLDIR/bin/$bin _stage/usr/local/ncbi/blast/bin
+    if [ $? -ne 0 ]; then
+        echo FAILURE
+        exit 1;
+    fi
 done
 
 echo building package
 mkdir $PRODUCT
 /Developer/usr/bin/packagemaker --id gov.nih.nlm.ncbi.blast --doc ncbi-blast.pmdoc --out $PRODUCT/$PRODUCT.pkg
+if [ $? -ne 0 ]; then
+    echo FAILURE
+    exit 1;
+fi
 
 echo copying uninstaller
 cp -p uninstall_ncbi_blast.zip $PRODUCT
+if [ $? -ne 0 ]; then
+    echo FAILURE
+    exit 1;
+fi
 
 echo creating disk image
 /usr/bin/hdiutil create $PRODUCT.dmg -srcfolder $PRODUCT
+if [ $? -ne 0 ]; then
+    echo FAILURE
+    exit 1;
+fi
 
 echo moving disk image
 mkdir $INSTALLDIR/installer
+if [ $? -ne 0 ]; then
+    echo FAILURE
+    exit 1;
+fi
 mv $PRODUCT.dmg $INSTALLDIR/installer
+if [ $? -ne 0 ]; then
+    echo FAILURE
+    exit 1;
+fi
 
 echo done
 rm -rf _stage $PRODUCT
